+++ openiboot/plat-s5l8900/ftl.c
@@ -3260,7 +3260,7 @@ void ftl_printdata() {
 
 }
 
-static error_t ftl_read_mtd(mtd_t *_dev, void *_dest, uint32_t _off, int _amt)
+static error_t ftl_read_mtd(mtd_t *_dev, void *_dest, uint64_t _off, int _amt)
 {
 	int ret = ftl_read(_dest, _off, _amt);
 	if(ret < 0)
@@ -3269,7 +3269,7 @@ static error_t ftl_read_mtd(mtd_t *_dev, void *_dest, uint32_t _off, int _amt)
 	return SUCCESS_VALUE(ret);
 }
 
-static error_t ftl_write_mtd(mtd_t *_dev, void *_src, uint32_t _off, int _amt)
+static error_t ftl_write_mtd(mtd_t *_dev, void *_src, uint64_t _off, int _amt)
 {
 	int ret = ftl_write(_src, _off, _amt);
 	if(ret < 0)
@@ -3278,7 +3278,7 @@ static error_t ftl_write_mtd(mtd_t *_dev, void *_src, uint32_t _off, int _amt)
 	return SUCCESS_VALUE(ret);
 }
 
-static int ftl_block_size(mtd_t *_dev)
+static int64_t ftl_block_size(mtd_t *_dev)
 {
 	NANDData* Data = nand_get_geometry();
 	return Data->bytesPerPage;
